users can now rate books

+++ include/book.h
@@ -8,17 +8,17 @@
 class Book
 {
 private:
-    int m_isbn;
+    long m_isbn;
     std::string m_description;
 
 public:
     Book();
-    Book(int isbn, std::string description);
+    Book(long isbn, std::string description);
 
-    void setIsbn(int isbn) {m_isbn = isbn;}
+    void setIsbn(long isbn) {m_isbn = isbn;}
     void setDescription(std::string des) {m_description = des;}
     
-    int getIsbn() {return m_isbn;}
+    long getIsbn() {return m_isbn;}
     std::string getDescription() {return m_description;}
     
     friend std::istream& operator>>(std::istream& in, Book& book);
