+++ src/book.cpp
@@ -9,7 +9,7 @@ Book::Book()
     m_description = "";
 }
 
-Book::Book(int isbn, string description)
+Book::Book(long isbn, string description)
 {
     m_isbn = isbn;
     m_description = description;
@@ -20,7 +20,7 @@ istream& operator>>(istream& in, Book& book)
     string temp = "";
     getline(in, temp);
     int position = temp.find(',');
-    book.m_isbn = stoi(temp.substr(0, position));
+    book.m_isbn = stol(temp.substr(0, position));
     book.m_description = temp.substr(position + 2, temp.length()); 
     return in;
 }
